api-types, main, vmm: move `VmParams` to api-types
Part of the move to an api-types crate.

On-behalf-of: SAP julian.schindel@sap.com
